requirements.txt
Django>=5.2.4
google-generativeai>=0.8.0
python-dotenv>=1.0.0
python-docx>=1.1.0

